# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get pending applications for the organizer",
                "description": "Get pending applications for the organizer.\n\nRetrieves pending applications for the authenticated organizer.",
                "operationId": "5bdb0276efb6e85e52a842b69a9c6601",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "pending_applications": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Application"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Applications"
                ],
                "summary": "Store a new application",
                "description": "Store a new application.\n\nAllows a volunteer to apply to a volunteer initiative.",
                "operationId": "838ab9ccbecd19fc861d7a1e17c865ba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "announcement_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "volunteer_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Application submitted successfully."
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications/{application}": {
            "patch": {
                "tags": [
                    "Applications"
                ],
                "summary": "Manage applications for the organizer",
                "description": "Manage applications for the organizer.\n\nAllows the organizer to approve or reject applications from volunteers.",
                "operationId": "62a850e43bad4c5c012b46f6d5e59e21",
                "parameters": [
                    {
                        "name": "application",
                        "in": "path",
                        "description": "ID of the application",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Application approved successfully."
                                        },
                                        "application": {
                                            "$ref": "#/components/schemas/Application"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Application": {
                "title": "Application",
                "description": "Application schema",
                "properties": {
                    "id": {
                        "description": "ID of the application",
                        "type": "integer"
                    },
                    "volunteer_id": {
                        "description": "ID of the volunteer",
                        "type": "integer"
                    },
                    "announcement_id": {
                        "description": "ID of the announcement",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}